# arm: reorganize files, optimize memcpy, memset