--- a/arch/arm/v7a/Makefile
+++ b/arch/arm/v7a/Makefile
@@ -1,0 +1,8 @@
+#
+# Makefile for libphoenix/arch/arm/v7a
+#
+# Copyright 2017, 2020 Phoenix Systems
+# Author: Pawel Pisarczyk
+#
+
+OBJS += $(addprefix $(PREFIX_O)arch/arm/v7a/, syscalls.o reboot.o)