--- a/arch/arm/v7m/Makefile
+++ b/arch/arm/v7m/Makefile
@@ -1,0 +1,8 @@
+#
+# Makefile for libphoenix/arch/arm/v7m
+#
+# Copyright 2017, 2020 Phoenix Systems
+# Author: Pawel Pisarczyk
+#
+
+OBJS += $(addprefix $(PREFIX_O)arch/arm/v7m/, syscalls.o reboot.o tls.o)