… arch/arm/crt0.S
… arch/arm/jmp.S
… arch/arm/v7a/reboot.c
… arch/arm/v7a/syscalls.S
… arch/arm/v7m/reboot.c
… arch/arm/v7m/syscalls.S
… arch/arm/v7m/tls.c
… arch/armv7a/crt0.S
… arch/armv7a/string.c
… arch/armv7m/jmp.S
… arch/armv7m/signal.S