<!--- Provide a general summary of your changes in the Title above -->

## Description
<!--- Describe your changes shortly -->
Reorganize ARM files:
- create common `arm` directory
- extract common arm routines
- add `v7a`, `v7m` subdirectories with files specific for that arch

`memcpy` optimization:
- handle misaligned buffers,
- copy in parts - basic cases cover:
    - `len` $<$ 64
    -  64  $\leq$ `len`

`memset` optimization:
- code divided into 2 cases:
    - `len` $<$ 64
    -  64  $\leq$ `len`
### Memcpy benchmarks:

#### Old implementation

```
off          0 |        1 |        2 |        3 |        4 |        5 |        6 |        7 |        8 | 
--------------------------------------------------------------------------------------------------------
b8:        235 |      256 |      222 |      223 |      108 |      254 |      222 |      249 |      108 | 
b16:       167 |      412 |      442 |      435 |      164 |      438 |      439 |      409 |      167 | 
b32:       288 |      833 |      824 |      822 |      291 |      794 |      794 |      818 |      310 | 
b64:       588 |     1630 |     1594 |     1580 |      571 |     1604 |     1606 |     1586 |      547 | 
b128:     1040 |     3164 |     3178 |     3187 |     1013 |     3152 |     3168 |     3186 |     1039 | 
b256:     1928 |     6343 |     6355 |     6353 |     1916 |     6387 |     6381 |     6375 |     1909 | 
b512:     3685 |    12666 |    12681 |    12660 |     3668 |    12645 |    12664 |    12724 |     3709 | 
b1024:    7671 |    26407 |    26490 |    26427 |     7662 |    26550 |    26528 |    26549 |     7648 | 
b2048:   14812 |    51817 |    51801 |    51852 |    14779 |    51865 |    51663 |    51932 |    14777 | 
b4096:   29007 |   102689 |   102538 |   102299 |    29026 |   102677 |   102962 |   102688 |    29013 | 
b8192:   57764 |   204702 |   205158 |   204743 |    57673 |   205559 |   205738 |   204963 |    57732 | 
b16384  115163 |   414076 |   413271 |   412504 |   114849 |   413162 |   413003 |   413334 |   114911 | 
b32768  229685 |   830554 |   828351 |   828070 |   229596 |   831787 |   829131 |   828910 |   229745 | 
```
`off` is offset from aligned address, first column contains number of bytes copied.
Times are a sum of 10000 iterations in us.


#### New implementation
```
off          0 |        1 |        2 |        3 |        4 |        5 |        6 |        7 |        8 | 
--------------------------------------------------------------------------------------------------------
b8:        251 |      213 |      238 |      212 |      211 |      214 |      211 |      231 |      240 | 
b16:       255 |      252 |      251 |      281 |      283 |      254 |      260 |      255 |      251 | 
b32:       341 |      369 |      337 |      338 |      341 |      341 |      359 |      343 |      330 | 
b64:       544 |      606 |      585 |      582 |      551 |      522 |      501 |      524 |      476 | 
b128:      680 |      790 |      770 |      742 |      716 |      712 |      697 |      672 |      626 | 
b256:     1065 |     1138 |     1119 |     1150 |     1089 |     1063 |     1073 |     1075 |      997 | 
b512:     2321 |     1853 |     1839 |     1808 |     1812 |     2217 |     2182 |     2162 |     2163 | 
b1024:    4298 |     4555 |     4542 |     4615 |     4613 |     4586 |     4587 |     4503 |     4512 | 
b2048:    7757 |     7376 |     7367 |     7578 |     7436 |     7427 |     7488 |     7428 |     7446 | 
b4096:   14355 |    14003 |    13998 |    14031 |    14089 |    13805 |    13730 |    13956 |    13695 | 
b8192:   26073 |    25767 |    25790 |    25832 |    25984 |    25158 |    25239 |    25148 |    25266 | 
b16384   50478 |    49692 |    49739 |    49582 |    50499 |    48158 |    48262 |    48830 |    48462 | 
b32768   99926 |    98652 |    98188 |   100087 |    97771 |    94726 |    94758 |    95663 |    96211 | 
```

About 2 - 2,5x speed up with aligned addresses and up to 8,5x with unaligned access.

## Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->

## Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

<!--- In case of breaking change - please advice here what needs to be done in dependent projects. -->

## How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
- [x] Already covered by automatic testing.
- [ ] New test added: (add PR link here).
- [x] Tested by hand on: `armv7a9-zynq7000-qemu`

## Checklist:
<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [ ] I have added tests to cover my changes.
- [x] All new and existing linter checks and tests passed.
- [x] My changes generate no new compilation warnings for any of the targets.

## Special treatment

- [ ] This PR needs additional PRs to work (list the PRs, preferably in merge-order).
- [ ] I will merge this PR by myself when appropriate.