@nalajcie 2024-02-27
- maybe check `len` not divisible by 2? (probably more common case than unaligned beginning of the buffer)
 - `off` from the aligned address regards `src`/`dst`, or diff between them?

Maybe we can benchmark it against naive C-only implementation with loop unrolling - just for reference? - eg. newlib one
(https://github.com/bminor/newlib/blob/master/newlib/libc/string/memcpy.c)?

Did you try benchmarking it against eg. uClibc-ng arm asm optimized version? (https://github.com/wbx-github/uclibc-ng/blob/master/libc/string/arm/_memcpy.S)