@lukileczo 2024-03-01
@nalajcie regarding your comment - len divisibility (by 2, 4, whatever) does not meaningfully impact performance, as it's always handled in the tail. `off` in the benchmarks is the same value for both src and dst `memcpy(src + off, dst + off, len)`. Lower I attached benchmarks, where src and dst aren't mutually aligned (however there is no visible performance impact).

Some more benchmarks:
## New `memcpy`
### `.arm`
#### src/dst mutually aligned
```
off          0 |        1 | 
------------------------------
b8:        247 |      208 | 
b16:       248 |      248 | 
b32:       368 |      358 | 
b64:       414 |      551 | 
b128:      594 |      737 | 
b256:      929 |     1087 | 
b512:     2185 |     1935 | 
b1024:    3510 |     3754 | 
b2048:    6284 |     7016 | 
b4096:   12579 |    13195 | 
b8192:   25140 |    25711 | 
b16384   50216 |    51356 | 
b32768   98060 |   101563 | 
```
#### dst misaligned by 1 - `memcpy(src + off, dst + off + 1, len)`
```
off          0 |        1 | 
------------------------------
b8:        231 |      207 | 
b16:       320 |      263 | 
b32:       321 |      330 | 
b64:       653 |      617 | 
b128:      774 |      774 | 
b256:     1202 |     1184 | 
b512:     1967 |     1984 | 
b1024:    3367 |     3425 | 
b2048:    6320 |     6674 | 
b4096:   12178 |    12107 | 
b8192:   24066 |    24157 | 
b16384   47823 |    48275 | 
b32768   94652 |    99616 |
```
#### src misaligned by 1 - `memcpy(src + off + 1, dst + off, len)`
```
off          0 |        1 | 
------------------------------
b8:        327 |      237 | 
b16:       248 |      247 | 
b32:       327 |      323 | 
b64:       402 |      613 | 
b128:      542 |      805 | 
b256:      920 |     1149 | 
b512:     1813 |     1957 | 
b1024:    3450 |     3367 | 
b2048:    6027 |     6387 | 
b4096:   11918 |    12246 | 
b8192:   23796 |    24361 | 
b16384   47536 |    47897 | 
b32768   94394 |    94629 | 
```

### `.thumb`
#### src/dst mutually aligned
This is a little bit concerning for me as the performance of aligned copy is lower than unaligned - but manual loop alignment stays the same (32 bytes, equal icache line) so I don't really know why it's happening.

```
off          0 |        1 | 
------------------------------
b8:        248 |      207 | 
b16:       246 |      246 | 
b32:       325 |      361 | 
b64:       441 |      537 | 
b128:      604 |      724 | 
b256:      890 |     1101 | 
b512:     2515 |     1891 | 
b1024:    4386 |     4455 | 
b2048:    8489 |     8667 | 
b4096:   16865 |    16790 | 
b8192:   33021 |    32957 | 
b16384   64768 |    64599 | 
b32768  129154 |   128443 | 
```
#### dst misaligned by 1
```
off          0 |        1 | 
------------------------------
b8:        224 |      203 | 
b16:       246 |      263 | 
b32:       406 |      321 | 
b64:       611 |      606 | 
b128:      817 |      770 | 
b256:     1174 |     1174 | 
b512:     1996 |     1945 | 
b1024:    3485 |     3370 | 
b2048:    6345 |     6316 | 
b4096:   12284 |    12151 | 
b8192:   24209 |    23919 | 
b16384   48020 |    47358 | 
b32768   94502 |    94557 | 
```

#### src misaligned by 1 
```
off          0 |        1 | 
------------------------------
b8:        224 |      208 | 
b16:       242 |      247 | 
b32:       396 |      326 | 
b64:       522 |      798 | 
b128:      749 |      818 | 
b256:     1094 |     1368 | 
b512:     1906 |     2128 | 
b1024:    3308 |     3572 | 
b2048:    6256 |     6883 | 
b4096:   12146 |    13084 | 
b8192:   25058 |    24164 | 
b16384   53383 |    49236 | 
b32768  105324 |    96130 | 
```

## C memcpy
(https://github.com/bminor/newlib/blob/master/newlib/libc/string/memcpy.c)
```
off          0 |        1 | 
------------------------------
b8:        445 |      414 | 
b16:       346 |      545 | 
b32:       415 |     1024 | 
b64:       488 |     1493 | 
b128:      761 |     2770 | 
b256:     1257 |     5344 | 
b512:     2318 |    10380 | 
b1024:    4242 |    20532 | 
b2048:    8271 |    40605 | 
b4096:   16320 |    81429 | 
b8192:   32325 |   160967 | 
b16384   64264 |   320014 | 
b32768  130048 |   640968 | 
```

## uClibc - `.thumb`
(https://github.com/wbx-github/uclibc-ng/blob/master/libc/string/arm/_memcpy.S)

#### src/dst mutually aligned
```
off          0 |        1 | 
------------------------------
b8:        388 |      392 | 
b16:       305 |      349 | 
b32:       383 |      431 | 
b64:       481 |      528 | 
b128:      651 |      821 | 
b256:     1031 |     1099 | 
b512:     1767 |     1848 | 
b1024:    3336 |     3359 | 
b2048:    6431 |     6442 | 
b4096:   12597 |    12598 | 
b8192:   24757 |    24930 | 
b16384   49676 |    49730 | 
b32768  100916 |   101239 | 
```

#### dst misaligned by 1
```
off          0 |        1 | 
------------------------------
b8:        436 |      322 | 
b16:       369 |      370 | 
b32:       478 |      488 | 
b64:       593 |      592 | 
b128:      844 |      841 | 
b256:     1368 |     1348 | 
b512:     2506 |     2299 | 
b1024:    4489 |     4292 | 
b2048:    8738 |     8458 | 
b4096:   16536 |    16385 | 
b8192:   31989 |    32126 | 
b16384   63783 |    63841 | 
b32768  129096 |   128766 | 
```

#### src misaligned by 1 
```
off          0 |        1 | 
------------------------------
b8:        321 |      326 | 
b16:       413 |      394 | 
b32:       381 |      462 | 
b64:       552 |      608 | 
b128:      767 |      850 | 
b256:     1235 |     1471 | 
b512:     2273 |     2346 | 
b1024:    4153 |     4440 | 
b2048:    8363 |     8351 | 
b4096:   16380 |    16209 | 
b8192:   31899 |    32024 | 
b16384   63710 |    63781 | 
b32768  128189 |   128997 | 
```